Deprecate Atlas provider, and use the "Atlas" name to avoid future confusion
This provider was only used to read data from the artifact registry, which we
did not recreate in the new Terraform Enterprise. For most users, there's
nothing for it to talk to anymore, since the remaining Atlas code has stopped
running Packer builds and serving artifacts.

In many places in the docs, we changed "Atlas" to "Terraform Enterprise," since
Terraform was the biggest part of Atlas. However, this provider is one of the
few cases where the name update doesn't make sense: it only relates to a
non-Terraform part of the old Atlas suite. Since there will likely be a real
Terraform Enterprise provider in the future, we should go back to referring to
this provider by its old name. (This commit doesn't change the URLs yet; we can
move the files in a future edit.)